Polish LoggerUtils, prevent nested install of CapturingFactory (#2512)
This commit removes the internal `LoggerFactory` interface from the
LoggerUtils class, directly storing the factories as `Function`s.

This has the added benefit that the CapturingFactory can reflexively
detect another one is already in place (which was trickier before, as
the field would contain a method reference, not the whole instance).

As a consequence, attempting to install a CapturingFactory over another
CapturingFactory will result in the old one being returned.

This commit also contains a few polish changes (fixing typos, etc...).